fixed menu bug and seperated into .cpp and .h files
